docker-monitoring/requirements.txt
docker
requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
logging
datetime
certifi>=2024.7.4 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=70.0.0 # not directly required, pinned by Snyk to avoid a vulnerability
python-dotenv